[update] 优化代码

+++ src/main/java/cn/codeartist/code/generator/builder/FileBuilder.java
@@ -1,16 +1,13 @@
 package cn.codeartist.code.generator.builder;
 
+import cn.codeartist.code.generator.config.Configuration;
 import cn.codeartist.code.generator.exceptions.BuilderException;
 import cn.codeartist.code.generator.parsing.Template;
 import freemarker.cache.ClassTemplateLoader;
 import freemarker.cache.MruCacheStorage;
-import freemarker.template.Configuration;
 import org.apache.log4j.Logger;
 
-import java.io.BufferedWriter;
-import java.io.FileOutputStream;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
+import java.io.*;
 import java.nio.charset.StandardCharsets;
 
 /**
@@ -21,13 +18,13 @@ import java.nio.charset.StandardCharsets;
 public class FileBuilder {
 
     private final Logger logger = Logger.getLogger(FileBuilder.class);
-    private final Configuration configuration;
+    private final freemarker.template.Configuration config;
 
-    public FileBuilder() {
-        this.configuration = new Configuration(Configuration.VERSION_2_3_23);
-        this.configuration.setTemplateLoader(new ClassTemplateLoader(FileBuilder.class, "/template"));
-        this.configuration.setCacheStorage(new MruCacheStorage(10, 100));
-        this.configuration.setDefaultEncoding("UTF-8");
+    public FileBuilder(Configuration configuration) {
+        this.config = new freemarker.template.Configuration(freemarker.template.Configuration.VERSION_2_3_23);
+        this.config.setTemplateLoader(new ClassTemplateLoader(FileBuilder.class, configuration.getSettings().getTemplatePath()));
+        this.config.setCacheStorage(new MruCacheStorage(10, 100));
+        this.config.setDefaultEncoding("UTF-8");
     }
 
     public void build(Template template, String path, String className, Object object) {
@@ -36,7 +33,12 @@ public class FileBuilder {
 
     public void build(Template template, String path, Object object) {
         try {
-            freemarker.template.Template tpl = configuration.getTemplate(template.name + ".ftl");
+            freemarker.template.Template tpl = config.getTemplate(template.name + ".ftl");
+            // 只生成一次，不重复生成下面代码
+            boolean oneGenerate = template == Template.BASEDAO || template == Template.INTERFACE || template == Template.SERVICE || template == Template.CONTROLLER;
+            if (oneGenerate && new File(path).exists()) {
+                return;
+            }
             FileOutputStream fos = new FileOutputStream(path);
             Writer out = new BufferedWriter(new OutputStreamWriter(fos, StandardCharsets.UTF_8), 10240);
             logger.info(template.name + " --> " + path + " successful");
